feat: separate custom markdown syntax, improve markdown rendering
